+++ crates/swc_html_parser/src/parser/mod.rs
@@ -3304,8 +3304,7 @@ where
                     } if tag_name == "input" => {
                         let is_self_closing = *self_closing;
                         let input_type = attributes
-                            .clone()
-                            .into_iter()
+                            .iter()
                             .find(|attribute| attribute.name.as_ref() == "type");
                         let is_hidden = match &input_type {
                             Some(input_type) => match &input_type.value {
@@ -3318,15 +3317,17 @@ where
                         };
 
                         self.reconstruct_active_formatting_elements()?;
+
+                        // To avoid extra cloning, it doesn't have effect on logic
+                        if input_type.is_none() || !is_hidden {
+                            self.frameset_ok = false;
+                        }
+
                         self.insert_html_element(token_and_info)?;
                         self.open_elements_stack.pop();
 
                         if is_self_closing {
                             token_and_info.acknowledged = true;
-                        }
-
-                        if input_type.is_none() || !is_hidden {
-                            self.frameset_ok = false;
                         }
                     }
                     // A start tag whose tag name is one of: "param", "source", "track"
@@ -3682,9 +3683,7 @@ where
                     }
                     // Any other end tag
                     Token::EndTag { .. } => {
-                        self.any_other_end_tag_for_in_body_insertion_mode(
-                            &mut token_and_info.clone(),
-                        );
+                        self.any_other_end_tag_for_in_body_insertion_mode(token_and_info);
                     }
                 }
 
@@ -4187,8 +4186,7 @@ where
                     } if tag_name == "input" => {
                         let is_self_closing = *self_closing;
                         let input_type = attributes
-                            .clone()
-                            .into_iter()
+                            .iter()
                             .find(|attribute| attribute.name.as_ref() == "type");
                         let is_hidden = match &input_type {
                             Some(input_type) => match &input_type.value {
@@ -6216,7 +6214,7 @@ where
                             span: Default::default(),
                             namespace: None,
                             prefix: None,
-                            name: attribute_token.name.clone(),
+                            name: attribute_token.name,
                             value: attribute_token.value,
                         };
 
@@ -6570,7 +6568,7 @@ where
                 self.open_elements_stack
                     .replace(node_index, new_element.clone());
 
-                node = new_element.clone();
+                node = new_element;
 
                 // 13.7
                 if is_same_node(&last_node, &furthest_block.1) {
@@ -6578,10 +6576,10 @@ where
                 }
 
                 // 13.8
-                self.append_node(&node, last_node.clone());
+                self.append_node(&node, last_node);
 
                 // 13.9
-                last_node = node.clone();
+                last_node = node;
             }
 
             // 14.